Update demo app for release 4.3.0 (#931)
