+++ DemoApp/ios/Podfile
@@ -6,9 +6,9 @@ target 'DemoApp' do
   # use_frameworks!
 
   # Pods for DemoApp
-  pod 'AppCenter/Crashes', '~> 4.2.0'
-  pod 'AppCenter/Analytics', '~> 4.2.0'
-  pod 'AppCenterReactNativeShared', '~> 4.2.0'
+  pod 'AppCenter/Crashes', '~> 4.3.0'
+  pod 'AppCenter/Analytics', '~> 4.3.0'
+  pod 'AppCenterReactNativeShared', '~> 4.3.0'
 
   platform :ios, '9.0'
